wxConfig() ambiguity resolved
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@1511 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/forty/scorefil.cpp
@@ -175,9 +175,9 @@ void ScoreFile::WritePlayersScore(const char* player, int wins, int games, int s
 
 		m_config->SetPath("/Players");
 		m_config->SetPath(player);
-		m_config->Write("Score", score);
-		m_config->Write("Games", games);
-		m_config->Write("Wins", wins);
+		m_config->Write("Score", (long)score);
+		m_config->Write("Games", (long)games);
+		m_config->Write("Wins", (long)wins);
 		m_config->Write("Check", CalcCheck(player, games, wins, score));
 	}
 }
